Moving IntentsService into RoutesService and updating imports and tests
