+++ packages/sdk/src/index.ts
@@ -1,9 +1,6 @@
 export { RoutesService } from "./routes";
-export type { CreateSimpleRouteParams, CreateRouteParams } from "./routes/types";
+export type { CreateSimpleRouteParams, CreateRouteParams, SetupIntentForPublishingParams, IntentData } from "./routes/types";
 
 export { OpenQuotingClient } from "./quotes";
 export type { SolverQuote, QuoteData } from "./quotes/types";
 export { selectCheapestQuote } from "./quotes/quoteSelectors";
-
-export { IntentsService } from "./intents";
-export type { SetupIntentForPublishingParams, IntentData } from "./intents/types";
