+++ packages/sdk/src/routes/types.ts
@@ -1,5 +1,6 @@
 import { Hex } from "viem"
 import { ChainId } from "../constants/types"
+import { SolverQuote } from "../quotes/types"
 
 export type CreateSimpleRouteParams = {
   originChainID: ChainId
@@ -21,4 +22,21 @@ export type CreateRouteParams = {
   prover: "HyperProver" | "StorageProver" | Hex
   destinationChainActions: Hex[]
   expiryTime?: Date
+}
+
+export type SetupIntentForPublishingParams = {
+  creator: Hex
+  intentData: IntentData
+  quote: SolverQuote
+}
+
+export type IntentData = {
+  originChainID: ChainId
+  destinationChainID: ChainId
+  targetTokens: Hex[]
+  rewardTokens: Hex[]
+  rewardTokenBalances: bigint[]
+  proverContract: Hex
+  destinationChainActions: Hex[]
+  expiryTime: Date
 }
